requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==1.2.2
black==22.3.0
bleach==5.0.0
bootstrap-py==1.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.30
cytoolz==0.11.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.15  #HAD TO BE CHANGED
django-bootstrap-v5==1.0.11
django-rest-knox==4.2.0
djangorestframework==3.13.1
docutils==0.18.1
entrypoints==0.4
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.32.0
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
hexbytes==0.2.2
idna==3.3
imagesize==1.4.1
ipfshttpclient==0.8.0a2
ipykernel==6.13.0
ipython==7.34.0 #NEEDED TO CHANGE FROM 8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lru-dict==1.1.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
netaddr==0.8.0
numpy==1.21.6 #NEEDED TO CHANGE FROM 1.22.3
packaging==21.3
pandas==1.3.5 #NEEDED TO CHANGE FROM 1.4.2
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pguard==0.4.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
rlp==2.0.1
scikit-learn==1.0.2
scipy==1.7.3 #NEEDED TO CHANGE FROM 1.8.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
varint==1.0.2
wcwidth==0.2.5
web3==5.29.0
webencodings==0.5.1
websockets==9.1
yarl==1.7.2
